pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "garmin-dashboard"
version = "0.2.0"
description = "Export and analyze Garmin Connect runs/workouts with optional FIT metrics"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "garminconnect>=0.2.20",
  "garth>=0.4.47",
  "fitparse>=1.2.0",
  "python-dotenv>=1.0.0",
  "pandas>=2.0.0",
  "streamlit>=1.30.0"
]

[project.scripts]
garmin-dashboard = "garmin_dashboard.main:main"
garmin-dashboard-web = "garmin_dashboard.web_app:main"
